Move default channel factories to Behat-related services
