+++ src/Sylius/Behat/Service/DefaultChannelFactoryInterface.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Sylius\Component\Core\Test\Services;
+namespace Sylius\Behat\Service;
 
 /**
  * @author Mateusz Zalewski <mateusz.zalewski@lakion.com>
