+++ src/Sylius/Behat/spec/Service/DefaultChannelFactorySpec.php
@@ -9,13 +9,13 @@
  * file that was distributed with this source code.
  */
 
-namespace spec\Sylius\Component\Core\Test\Services;
+namespace spec\Sylius\Behat\Service;
 
 use PhpSpec\ObjectBehavior;
 use Sylius\Component\Channel\Factory\ChannelFactoryInterface;
 use Sylius\Component\Core\Model\ChannelInterface;
-use Sylius\Component\Core\Test\Services\DefaultChannelFactory;
-use Sylius\Component\Core\Test\Services\DefaultChannelFactoryInterface;
+use Sylius\Behat\Service\DefaultChannelFactory;
+use Sylius\Behat\Service\DefaultChannelFactoryInterface;
 use Sylius\Component\Currency\Model\CurrencyInterface;
 use Sylius\Component\Locale\Model\LocaleInterface;
 use Sylius\Component\Resource\Factory\FactoryInterface;
