+++ src/Sylius/Bundle/CoreBundle/Resources/config/test_services.xml
@@ -22,31 +22,6 @@
         <service id="sylius.behat.email_checker" class="Sylius\Component\Core\Test\Services\EmailChecker">
             <argument>%kernel.cache_dir%/spool</argument>
         </service>
-
-        <!--Default Store Data-->
-        <service id="sylius.behat.factory.default_united_states_channel" class="Sylius\Component\Core\Test\Services\DefaultUnitedStatesChannelFactory">
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.repository.country" />
-            <argument type="service" id="sylius.repository.currency" />
-            <argument type="service" id="sylius.repository.locale" />
-            <argument type="service" id="sylius.repository.zone" />
-            <argument type="service" id="sylius.factory.channel" />
-            <argument type="service" id="sylius.factory.country" />
-            <argument type="service" id="sylius.factory.currency" />
-            <argument type="service" id="sylius.factory.locale" />
-            <argument type="service" id="sylius.factory.zone" />
-            <argument>%locale%</argument>
-        </service>
-        <service id="sylius.behat.factory.default_channel" class="Sylius\Component\Core\Test\Services\DefaultChannelFactory">
-            <argument type="service" id="sylius.factory.channel" />
-            <argument type="service" id="sylius.factory.currency" />
-            <argument type="service" id="sylius.factory.locale" />
-            <argument type="service" id="sylius.repository.channel" />
-            <argument type="service" id="sylius.repository.currency" />
-            <argument type="service" id="sylius.repository.locale" />
-            <argument>%locale%</argument>
-        </service>
-
     </services>
 
 </container>
